# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgroSense API",
    "version": "0.1.0",
    "description": "Implemented development API. Production OTP, export/deletion and storage authorization endpoints are not yet implemented."
  },
  "servers": [
    {
      "url": "http://localhost:4100"
    }
  ],
  "paths": {
...
    "/crops": {
      "get": {
        "summary": "Catalogue and explicit live inference coverage",
        "operationId": "get_crops",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        }
      }
    },
    "/auth/dev": {
      "post": {
        "summary": "Development-only session; never phone verification",
        "operationId": "post_auth_dev",
        "security": [],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "enum": [
                      "farmer-demo",
                      "farmer-other",
                      "advisor-demo",
                      "admin-demo",
                      "reviewer-demo"
                    ]
                  }
                },
                "required": [
                  "user_id"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Revoke current session",
        "operationId": "post_auth_logout",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        }
      }
    },
    "/me": {
      "get": {
        "summary": "Current account",
        "operationId": "get_me",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        }
      }
    },
    "/fields": {
      "get": {
        "summary": "List own fields",
        "operationId": "get_fields",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Field"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        }
      },
      "post": {
        "summary": "Create owned field",
        "operationId": "post_fields",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Field"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldInput"
              }
            }
          }
        }
      }
    },
    "/cases": {
      "get": {
        "summary": "List own cases (symptoms and analysis are serialized JSON in list rows)",
        "operationId": "get_cases",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        }
      },
      "post": {
        "summary": "Idempotent case creation; ownership and crop validated",
        "operationId": "post_cases",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Case"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseInput"
              }
            }
          }
        }
      }
    },
    "/cases/{id}": {
      "get": {
        "summary": "Authorized full case",
        "operationId": "get_cases__id_",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Case"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/cases/{id}/images": {
      "post": {
        "summary": "Upload private JPEG/PNG, 8 MB limit; re-encoded and deduplicated by checksum",
        "operationId": "post_cases__id__images",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "image/jpeg": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "image/png": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      }
    },
    "/cases/{id}/images/{image}": {
      "get": {
        "summary": "Read authorized private photo",
        "operationId": "get_cases__id__images__image_",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "image",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/cases/{id}/analysis": {
      "post": {
        "summary": "Queue exactly one analysis job per case",
        "operationId": "post_cases__id__analysis",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/cases/{id}/retry": {
      "post": {
        "summary": "Retry failed analysis only",
        "operationId": "post_cases__id__retry",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/cases/{id}/review": {
      "post": {
        "summary": "Request advisor review",
        "operationId": "post_cases__id__review",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Case"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/cases/{id}/follow-ups": {
      "post": {
        "summary": "Record idempotent outcome observation",
        "operationId": "post_cases__id__follow_ups",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Case"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowUpInput"
              }
            }
          }
        }
      }
    },
    "/reminders": {
      "get": {
        "summary": "List own reminders",
        "operationId": "get_reminders",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        }
      },
      "post": {
        "summary": "Upsert own reminder by client ID",
        "operationId": "post_reminders",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderInput"
              }
            }
          }
        }
      }
    },
    "/advice": {
      "get": {
        "summary": "Published guidance; development content only in development mode",
        "operationId": "get_advice",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Advice"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        }
      }
    },
    "/advisor/cases": {
      "get": {
        "summary": "Assigned cases only",
        "operationId": "get_advisor_cases",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Session required"
          },
          "403": {
            "description": "Insufficient role"
          },
          "404": {
            "description": "Missing or unauthorized resource"
          },
          "409": {
            "description": "Invalid transition or conflicting idempotency key"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message"
        ],
        "additionalProperties": false
      },
      "FieldInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "crop": {
            "type": "string",
            "enum": [
              "tomato",
              "maize",
              "mahangu",
              "sorghum"
            ]
          },
          "region": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "production_type": {
            "enum": [
              "rain_fed",
              "irrigated"
            ]
          },
          "planting_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          }
        },
        "required": [
          "name",
          "crop",
          "region",
          "production_type"
        ],
        "additionalProperties": false
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "enum": [
                "farmer",
                "advisor",
                "admin",
                "reviewer"
              ]
            }
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "roles",
          "created_at"
        ],
        "additionalProperties": false
      },
...
    }
  }
}